requirements.txt
numpy
pillow
matplotlib
selenium
lxml
opencv-python # consider headless for prod